accounts/requirements.txt
fastapi==0.79.0
uvicorn==0.17.6
pypika-tortoise==0.1.6
tortoise-orm==0.19.2
asyncpg==0.26.0
aerich==0.6.3
bcrypt==3.2.0
